+++ controller/bindings/locale/locale.mli
@@ -1,5 +1,5 @@
 val get_lang : unit -> (string option) Lwt.t
-val set_lang : string -> bool Lwt.t
+val set_lang : string -> unit Lwt.t
 
 val get_keymap : unit -> (string option) Lwt.t
-val set_keymap : string -> bool Lwt.t
+val set_keymap : string -> unit Lwt.t