+++ controller/server/network.mli
@@ -1,25 +1,7 @@
 (** Initialize Network connectivity *)
-val init : systemd : Systemd.Manager.t
-  -> connman : Connman.Manager.t
+val init
+  : connman : Connman.Manager.t
   -> (unit,exn) Lwt_result.t
-
-module Internet : sig
-  (** Internet connectivity state *)
-  type state =
-    | Pending
-    | Connected
-    | NotConnected of string
-  [@@deriving sexp]
-
-  val is_connected : state -> bool
-
-  (** [get connman] starts a thread that checks Internet connectivity periodically and on changes to network (changes in ConnMan services).
-
-      The state of internet connectivity is made available via a signal.
-  *)
-  val get: Connman.Manager.t -> (state Lwt_react.S.t * unit Lwt.t) Lwt.t
-
-end
 
 module Interface : sig
   (** Network interface *)