+++ controller/server/view/common/signal_strength.mli
@@ -1,0 +1,3 @@
+val html
+  : int
+  -> [> Html_types.svg ] Tyxml.Html.elt