+++ controller/server/view/error_page.mli
@@ -1,0 +1,8 @@
+type params =
+  { message: string
+  ; request: string
+  }
+
+val html :
+  params
+  -> [> Html_types.html ] Tyxml.Html.elt