+++ controller/server/view/status_page.mli
@@ -1,0 +1,10 @@
+type params =
+  { health: string
+  ; update: string
+  ; rauc: string
+  ; interfaces: string
+  }
+
+val html :
+  params
+  -> [> Html_types.html ] Tyxml.Html.elt