+++ pkgs/breeze-contrast-cursor-theme/default.nix
@@ -7,22 +7,17 @@ stdenv.mkDerivation rec {
   version = "1.0";
   themeName = "Breeze_Contrast";
 
-  src = fetchzip {
-    url = "https://code.jpope.org/jpope/breeze_cursor_sources/raw/master/${name}.zip";
-    sha256 = "1l8ils82bq2hlsl8shkcirxfjgk0459hsf6zvjnk9zrav47y9vjk";
-    # See issue https://github.com/NixOS/nixpkgs/issues/38649
-    extraPostFetch = "chmod go-w $out";
-  };
+  src = ./theme;
 
   installPhase = ''
-    install -d $out/share/icons/${themeName}
-    cp -rf * $out/share/icons/${themeName}
+    install -d $out/share/icons
+    cp -r share/icons/${themeName} $out/share/icons/${themeName}
   '';
 
   meta = {
     description = "Breeze Contrast cursor theme";
     homepage = https://kver.wordpress.com/2015/01/09/curses-i-mean-cursors/;
-    license = stdenv.lib.licenses.gpl3;
+    license = stdenv.lib.licenses.gpl2;
     platforms = stdenv.lib.platforms.all;
   };
 }